Add ReproductionService and migrate
