--- Assets/Scripts/Services/BreedingService.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: c23f01c1117b47569e8bad547ef3ff69
-timeCreated: 1748778737
