--- Assets/Scripts/Services/EggService.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: aa06c4195455457c91af57c05090af1f
-timeCreated: 1748778821
